+++ src/lib/pages/account-details/index.tsx
@@ -11,7 +11,6 @@ import { useCallback, useEffect } from "react";
 
 import { AmpEvent, useTrack } from "lib/amplitude";
 import {
-  useConvertHexAddress,
   useInternalNavigate,
   useMoveConfig,
   useValidateAddress,
@@ -23,6 +22,7 @@ import { CustomIcon } from "lib/components/icon";
 import { Loading } from "lib/components/Loading";
 import PageContainer from "lib/components/PageContainer";
 import { InvalidState } from "lib/components/state";
+import { useFormatAddresses } from "lib/hooks/useFormatAddresses";
 import { useAccountDetailsTableCounts } from "lib/model/account";
 import { useAccountId } from "lib/services/accountService";
 import type { IndexedModule } from "lib/services/move/moduleService";
@@ -35,18 +35,17 @@ import {
 } from "lib/services/publicProjectService";
 import type { HexAddr, HumanAddr, MoveAccountAddr, Option } from "lib/types";
 import {
-  bech32AddressToHex,
   getFirstQueryParam,
-  isHexAddress,
+  isHexWalletAddress,
+  isHexModuleAddress,
   truncate,
-  unpadHexAddress,
 } from "lib/utils";
 
 import { AccountHeader } from "./components/AccountHeader";
 import { AssetsSection } from "./components/asset";
 import { DelegationsSection } from "./components/delegations";
-import { ModuleLists } from "./components/ModuleLists";
-import { ResourceLists, ResourceSection } from "./components/resources";
+import { ModuleLists } from "./components/modules";
+import { ResourceOverview, ResourceSection } from "./components/resources";
 import {
   AdminContractsTable,
   InstantiatedContractsTable,
@@ -94,6 +93,16 @@ const AccountDetailsBody = ({
   const { data: accountId } = useAccountId(accountAddress);
   const { data: icnsName } = useICNSNamesByAddress(accountAddress);
 
+  const publicDetail = publicInfoBySlug?.details;
+
+  const {
+    tableCounts,
+    refetchCodesCount,
+    refetchContractsAdminCount,
+    refetchContractsCount,
+    refetchProposalsCount,
+    loadingState: { txCountLoading },
+  } = useAccountDetailsTableCounts(accountAddress, accountId);
   // TODO: combine with useAccountDetailsTableCounts and remove type assertion
   // move
   const { data: fetchedAccountModules, isFetching: isModulesLoading } =
@@ -109,16 +118,6 @@ const AccountDetailsBody = ({
     useAccountResources({
       address: accountAddress,
     });
-
-  const publicDetail = publicInfoBySlug?.details;
-  const {
-    tableCounts,
-    refetchCodesCount,
-    refetchContractsAdminCount,
-    refetchContractsCount,
-    refetchProposalsCount,
-    loadingState: { txCountLoading },
-  } = useAccountDetailsTableCounts(accountAddress, accountId);
 
   const handleTabChange = useCallback(
     (nextTab: TabIndex) => () => {
@@ -337,13 +336,16 @@ const AccountDetailsBody = ({
             )}
             {move.enabled && (
               <>
-                <ResourceLists
+                <ResourceOverview
                   address={accountAddress}
+                  totalCount={resourcesData?.totalCount}
+                  resourcesByName={resourcesData?.groupedByName}
+                  isLoading={isResourceLoading}
                   onViewMore={handleTabChange(TabIndex.Resources)}
                 />
                 <ModuleLists
-                  selectedAddress={accountAddress}
                   totalCount={modulesData?.length}
+                  selectedAddress={accountAddress}
                   modules={modulesData}
                   isLoading={isModulesLoading}
                   onViewMore={handleTabChange(TabIndex.Modules)}
@@ -395,7 +397,7 @@ const AccountDetailsBody = ({
           <TabPanel p={0}>
             <ResourceSection
               address={accountAddress}
-              resources={resourcesData?.groupedByOwner}
+              resourcesByOwner={resourcesData?.groupedByOwner}
               isLoading={isResourceLoading}
             />
           </TabPanel>
@@ -425,20 +427,12 @@ const AccountDetails = () => {
   const router = useRouter();
   const { track } = useTrack();
   const { validateUserAddress, validateContractAddress } = useValidateAddress();
-  const hexToBech32 = useConvertHexAddress();
+  const formatAddresses = useFormatAddresses();
   // TODO: change to `Addr` for correctness (i.e. interchain account)
   const accountAddressParam = getFirstQueryParam(
     router.query.accountAddress
   ).toLowerCase() as MoveAccountAddr;
-  const [hexAddr, humanAddr] = isHexAddress(accountAddressParam)
-    ? [
-        unpadHexAddress(accountAddressParam as HexAddr),
-        hexToBech32(accountAddressParam as HexAddr),
-      ]
-    : [
-        unpadHexAddress(bech32AddressToHex(accountAddressParam as HumanAddr)),
-        accountAddressParam as HumanAddr,
-      ];
+  const { address, hex } = formatAddresses(accountAddressParam);
   // TODO: fix assertion later
   const tab = getFirstQueryParam(router.query.tab) as TabIndex;
 
@@ -450,12 +444,13 @@ const AccountDetails = () => {
 
   return (
     <PageContainer>
-      {!isHexAddress(accountAddressParam) &&
+      {!isHexWalletAddress(accountAddressParam) &&
+      !isHexModuleAddress(accountAddressParam) &&
       validateUserAddress(accountAddressParam) &&
       validateContractAddress(accountAddressParam) ? (
         <InvalidAccount />
       ) : (
-        <AccountDetailsBody hexAddress={hexAddr} accountAddress={humanAddr} />
+        <AccountDetailsBody hexAddress={hex} accountAddress={address} />
       )}
     </PageContainer>
   );