+++ src/lib/pages/public-project/components/tables/modules/index.tsx
@@ -97,7 +97,7 @@ export const PublicProjectModuleTable = ({
         <EmptyState
           message={
             modules.length
-              ? "No matching module found for this project. Make sure you are searching with Account Address or Account Name"
+              ? "No matching module found for this project. Make sure you are searching with Module Address or Module Name"
               : "There is currently no modules related to this project."
           }
           imageVariant={onViewMore && "empty"}