+++ src/lib/services/searchService.ts
@@ -9,7 +9,12 @@ import {
   useGetAddressType,
 } from "lib/app-provider";
 import type { Addr, ContractAddr, Option } from "lib/types";
-import { isCodeId, isHexAddress, splitModule } from "lib/utils";
+import {
+  isCodeId,
+  isHexWalletAddress,
+  isHexModuleAddress,
+  splitModule,
+} from "lib/utils";
 
 import { useBlockInfoQuery } from "./blockService";
 import { useCodeDataByCodeId } from "./codeService";
@@ -65,11 +70,8 @@ export const useSearchHandler = (
   } = useCurrentChain();
   const getAddressType = useGetAddressType();
   const addressType = getAddressType(debouncedKeyword);
-  const { data: txData, isFetching: txFetching } = useTxData(debouncedKeyword);
-  const { data: codeData, isFetching: codeFetching } = useCodeDataByCodeId({
-    codeId: debouncedKeyword,
-    enabled: isWasm && isCodeId(debouncedKeyword),
-  });
+
+  // Contract
   const { data: contractData, isFetching: contractFetching } = useQuery(
     [CELATONE_QUERY_KEYS.CONTRACT_INFO, lcdEndpoint, debouncedKeyword],
     async () => queryContract(lcdEndpoint, debouncedKeyword as ContractAddr),
@@ -79,47 +81,17 @@ export const useSearchHandler = (
       retry: false,
     }
   );
-  const { data: blockData, isFetching: blockFetching } =
-    useBlockInfoQuery(debouncedKeyword);
-  const { data: poolData, isFetching: poolFetching } = usePoolByPoolId(
-    Number(debouncedKeyword),
-    isPool &&
-      Number.isInteger(Number(debouncedKeyword)) &&
-      Number(debouncedKeyword) > 0
-  );
-  const { data: icnsAddressData, isFetching: icnsAddressFetching } =
-    useAddressByICNSName(debouncedKeyword);
-
-  const [initiaAddr, moduleName, functionName] = splitModule(debouncedKeyword);
-
-  const enableModuleFetching = useMemo(
-    () =>
-      Boolean(
-        isMove &&
-          (getAddressType(initiaAddr) === "user_address" ||
-            isHexAddress(initiaAddr)) &&
-          moduleName &&
-          functionName === undefined
-      ),
-    [functionName, getAddressType, initiaAddr, isMove, moduleName]
-  );
-
-  const { data: moduleData, isFetching: moduleFetching } = useAccountModules({
-    address: initiaAddr,
-    moduleName,
-    functionName: undefined,
-    options: {
-      enabled: enableModuleFetching,
-      refetchOnWindowFocus: false,
-      retry: false,
-    },
-  });
 
   const isAddr =
     addressType === "user_address" ||
     addressType === "contract_address" ||
-    isHexAddress(debouncedKeyword);
+    (isMove &&
+      (isHexWalletAddress(debouncedKeyword) ||
+        isHexModuleAddress(debouncedKeyword)));
 
+  // ICNS
+  const { data: icnsAddressData, isFetching: icnsAddressFetching } =
+    useAddressByICNSName(debouncedKeyword);
   // provide ICNS metadata result
   const { data: icnsNames } = useICNSNamesByAddress(
     (isAddr ? debouncedKeyword : icnsAddressData?.address) as Addr
@@ -137,6 +109,56 @@ export const useSearchHandler = (
     );
   }, [isAddr, contractData, icnsAddressData]);
 
+  // Code
+  const { data: codeData, isFetching: codeFetching } = useCodeDataByCodeId({
+    codeId: debouncedKeyword,
+    enabled: isWasm && isCodeId(debouncedKeyword),
+  });
+
+  // Tx
+  const { data: txData, isFetching: txFetching } = useTxData(debouncedKeyword);
+
+  // Block
+  const { data: blockData, isFetching: blockFetching } =
+    useBlockInfoQuery(debouncedKeyword);
+
+  // Pool
+  const { data: poolData, isFetching: poolFetching } = usePoolByPoolId(
+    Number(debouncedKeyword),
+    isPool &&
+      !debouncedKeyword.startsWith("0x") &&
+      Number.isInteger(Number(debouncedKeyword)) &&
+      Number(debouncedKeyword) > 0
+  );
+
+  // Move
+  const [addr, moduleName, functionName] = splitModule(debouncedKeyword);
+
+  const enableModuleFetching = useMemo(
+    () =>
+      Boolean(
+        isMove &&
+          (getAddressType(addr) === "user_address" ||
+            isHexWalletAddress(addr)) &&
+          moduleName &&
+          functionName === undefined
+      ),
+    [functionName, getAddressType, addr, isMove, moduleName]
+  );
+
+  const { data: moduleData, isFetching: moduleFetching } = useAccountModules({
+    address: addr,
+    moduleName,
+    functionName: undefined,
+    options: {
+      enabled: enableModuleFetching,
+      refetchOnWindowFocus: false,
+      retry: false,
+    },
+  });
+
+  // TODO: handle module function later
+
   useEffect(() => {
     const timeoutId = setTimeout(() => {
       setDebouncedKeyword(keyword);
@@ -148,7 +170,7 @@ export const useSearchHandler = (
   return {
     results: [
       addressResult,
-      moduleData && "Module Path",
+      moduleName && moduleData && "Module Path",
       txData && "Transaction Hash",
       codeData && "Code ID",
       blockData && "Block",