+++ src/lib/types/staking.ts
@@ -1,0 +1,5 @@
+export interface StakingShare {
+  // string of a decimal
+  amount: string;
+  denom: string;
+}